# Upgraded Prometheus to further optimize ingester Push() on limits exceeded