--- a/pkg/ruler/compat.go
+++ b/pkg/ruler/compat.go
@@ -2,11 +2,13 @@ package ruler
 
 import (
 	"context"
+	"errors"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/prometheus/notifier"
+	"github.com/prometheus/prometheus/pkg/exemplar"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/pkg/value"
 	"github.com/prometheus/prometheus/promql"
@@ -48,6 +50,10 @@ func (a *pusherAppender) Append(_ uint64, l labels.Labels, t int64, v float64) (
 		Value:       v,
 	})
 	return 0, nil
+}
+
+func (a *pusherAppender) AppendExemplar(_ uint64, _ labels.Labels, _ exemplar.Exemplar) (uint64, error) {
+	return 0, errors.New("exemplars are unsupported")
 }
 
 func (a *pusherAppender) Commit() error {